requirements.txt
absl-py==2.1.0
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
contourpy==1.3.0
cryptography==43.0.1
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
docutils==0.21.2
flatbuffers==24.3.25
fonttools==4.54.1
idna==3.10
importlib_metadata==8.5.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.0.2
jax==0.4.33
jaxlib==0.4.33
jeepney==0.8.0
keyring==25.4.1
kiwisolver==1.4.7
markdown-it-py==3.0.0
matplot==0.1.9
matplotlib==3.9.2
mdurl==0.1.2
mediapipe==0.10.15
ml_dtypes==0.5.0
more-itertools==10.5.0
nh3==0.2.18
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
pillow==10.4.0
pkginfo==1.10.0
protobuf==4.25.5
pycparser==2.22
Pygments==2.18.0
pyloco==0.0.139
pyparsing==3.1.4
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
pyserial==3.5
python-dateutil==2.9.0.post0
readme_renderer==44.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.8.1
scipy==1.14.1
SecretStorage==3.3.3
SimpleWebSocketServer==0.1.2
six==1.16.0
sounddevice==0.5.0
twine==5.1.1
typing==3.7.4.3
urllib3==2.2.3
ushlex==0.99.1
websocket-client==1.8.0
zipp==3.20.2
